+++ apps/dashboard/shared/components/design-system/table/filters/CategoriesFilter.tsx
@@ -43,7 +43,6 @@ export const CategoriesFilter = ({
     <Popover open={isOpen} onOpenChange={setIsOpen}>
       <PopoverTrigger asChild>
         <ButtonFilter
-          onClick={() => setIsOpen(!isOpen)}
           isOpen={isOpen}
           hasFilters={hasFilters}
           className={className}